Add avoidOptionals.(query|mutation|subscription) options
